# src/main/java/org/esfe/dtos/categoria/CategoriaGuardar.java
package org.esfe.dtos.categoria;

import lombok.Getter;
import lombok.Setter;

import java.io.Serializable;

@Getter
@Setter
public class CategoriaGuardar implements Serializable {
    private String nombre;
    private String descripcion;
}