# swagger/MarketLocalShirtsAPI.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MarketLocalShirts API",
    "version": "1.0.0",
    "description": "API REST de MarketLocalShirts. Catálogo de camisas, autenticación JWT, usuarios y pedidos.\n\n1. POST /api/auth/login\n2. Copie el token\n3. Pulse Authorize y péguelo (sin la palabra Bearer)\n4. Pruebe el resto de endpoints."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Servidor local de desarrollo"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT de POST /api/auth/login. Pegue solo el token."
      }
    },
    "parameters": {
      "page": {
        "name": "page",
        "in": "query",
        "schema": { "type": "integer", "default": 0 }
      },
      "size": {
        "name": "size",
        "in": "query",
        "schema": { "type": "integer", "default": 20 }
      },
      "id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      }
    },
    "schemas": {
      "LoginGuardar": {
        "type": "object",
        "properties": {
          "correo": { "type": "string", "example": "admin@marketlocalshirts.com" },
          "contrasena": { "type": "string", "example": "Admin123" }
        }
      },
      "LoginSalida": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "tipoToken": { "type": "string", "example": "Bearer" },
          "usuario": { "$ref": "#/components/schemas/UsuarioSalida" }
        }
      },
      "UsuarioGuardar": {
        "type": "object",
        "properties": {
          "nombre": { "type": "string", "example": "Valeria Admin" },
          "correo": { "type": "string", "example": "admin@marketlocalshirts.com" },
          "telefono": { "type": "string", "example": "7777-7777" },
          "contrasena": { "type": "string", "example": "Admin123" },
          "rol": { "type": "string", "example": "CLIENTE" }
        }
      },
...
      "UsuarioSalida": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "nombre": { "type": "string" },
          "correo": { "type": "string" },
          "telefono": { "type": "string" },
          "rol": { "type": "string" },
          "activo": { "type": "boolean" }
        }
      },
      "CategoriaGuardar": {
        "type": "object",
        "properties": {
          "nombre": { "type": "string", "example": "Casual" },
          "descripcion": { "type": "string", "example": "Camisas de uso diario" }
        }
      },
      "CategoriaModificar": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "nombre": { "type": "string" },
          "descripcion": { "type": "string" }
        }
      },
      "CategoriaSalida": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "nombre": { "type": "string" },
          "descripcion": { "type": "string" }
        }
      },
      "MarcaGuardar": {
        "type": "object",
        "properties": {
          "nombre": { "type": "string", "example": "LocalWear" },
          "descripcion": { "type": "string" }
        }
      },
      "MarcaModificar": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "nombre": { "type": "string" },
          "descripcion": { "type": "string" }
        }
      },
      "MarcaSalida": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "nombre": { "type": "string" },
          "descripcion": { "type": "string" }
        }
      },
      "CamisaGuardar": {
        "type": "object",
        "properties": {
          "nombre": { "type": "string", "example": "Camisa Oxford" },
          "descripcion": { "type": "string", "example": "Manga larga" },
          "precio": { "type": "number", "example": 24.99 },
          "stock": { "type": "integer", "example": 10 },
          "imagenUrl": { "type": "string" },
          "talla": { "type": "string", "example": "M" },
          "color": { "type": "string", "example": "Blanco" },
          "categoriaId": { "type": "integer", "example": 1 },
          "marcaId": { "type": "integer", "example": 1 }
        }
      },
      "CamisaModificar": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "nombre": { "type": "string" },
          "descripcion": { "type": "string" },
          "precio": { "type": "number" },
          "stock": { "type": "integer" },
          "imagenUrl": { "type": "string" },
          "talla": { "type": "string" },
          "color": { "type": "string" },
          "categoriaId": { "type": "integer" },
          "marcaId": { "type": "integer" }
        }
      },
      "CamisaSalida": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "nombre": { "type": "string" },
          "descripcion": { "type": "string" },
          "precio": { "type": "number" },
          "stock": { "type": "integer" },
          "imagenUrl": { "type": "string" },
          "talla": { "type": "string" },
          "color": { "type": "string" },
          "categoria": { "$ref": "#/components/schemas/CategoriaSalida" },
          "marca": { "$ref": "#/components/schemas/MarcaSalida" }
        }
      },
...
    }
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Iniciar sesión",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginGuardar" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sesión iniciada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginSalida" }
              }
            }
          },
          "401": { "description": "Credenciales inválidas o cuenta inactiva" }
        }
      }
    },
    "/api/auth/registrar": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Registrar usuario",
        "description": "El primer usuario queda como ADMIN; los siguientes como CLIENTE.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UsuarioGuardar" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UsuarioSalida" }
              }
            }
          },
          "409": { "description": "El correo ya existe" }
        }
      }
    },
    "/api/auth/perfil": {
      "get": {
        "tags": ["Autenticación"],
        "summary": "Ver perfil autenticado",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Perfil",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UsuarioSalida" }
              }
            }
          },
          "401": { "description": "Sin token" }
        }
      }
    },
    "/api/categorias": {
      "get": {
        "tags": ["Categorías"],
        "summary": "Listar categorías",
        "responses": {
          "200": {
            "description": "Lista",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CategoriaSalida" }
                }
              }
            }
          },
          "204": { "description": "Sin datos" }
        }
      },
      "post": {
        "tags": ["Categorías"],
        "summary": "Crear categoría",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CategoriaGuardar" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CategoriaSalida" }
              }
            }
          }
        }
      }
    },
    "/api/categorias/paginado": {
      "get": {
        "tags": ["Categorías"],
        "summary": "Listar categorías paginadas",
        "parameters": [
          { "$ref": "#/components/parameters/page" },
          { "$ref": "#/components/parameters/size" }
        ],
        "responses": {
          "200": { "description": "Página" },
          "204": { "description": "Sin datos" }
        }
      }
    },
    "/api/categorias/{id}": {
      "get": {
        "tags": ["Categorías"],
        "summary": "Obtener categoría por ID",
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "responses": {
          "200": {
            "description": "Encontrada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CategoriaSalida" }
              }
            }
          },
          "404": { "description": "No existe" }
        }
      },
      "put": {
        "tags": ["Categorías"],
        "summary": "Editar categoría",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CategoriaModificar" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Actualizada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CategoriaSalida" }
              }
            }
          },
          "404": { "description": "No existe" }
        }
      },
      "delete": {
        "tags": ["Categorías"],
        "summary": "Eliminar categoría",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "responses": {
          "200": { "description": "Eliminada" }
        }
      }
    },
    "/api/marcas": {
      "get": {
        "tags": ["Marcas"],
        "summary": "Listar marcas",
        "responses": {
          "200": {
            "description": "Lista",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/MarcaSalida" }
                }
              }
            }
          },
          "204": { "description": "Sin datos" }
        }
      },
      "post": {
        "tags": ["Marcas"],
        "summary": "Crear marca",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MarcaGuardar" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MarcaSalida" }
              }
            }
          }
        }
      }
    },
    "/api/marcas/paginado": {
      "get": {
        "tags": ["Marcas"],
        "summary": "Listar marcas paginadas",
        "parameters": [
          { "$ref": "#/components/parameters/page" },
          { "$ref": "#/components/parameters/size" }
        ],
        "responses": {
          "200": { "description": "Página" },
          "204": { "description": "Sin datos" }
        }
      }
    },
    "/api/marcas/{id}": {
      "get": {
        "tags": ["Marcas"],
        "summary": "Obtener marca por ID",
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "responses": {
          "200": {
            "description": "Encontrada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MarcaSalida" }
              }
            }
          },
          "404": { "description": "No existe" }
        }
      },
      "put": {
        "tags": ["Marcas"],
        "summary": "Editar marca",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MarcaModificar" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Actualizada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MarcaSalida" }
              }
            }
          },
          "404": { "description": "No existe" }
        }
      },
      "delete": {
        "tags": ["Marcas"],
        "summary": "Eliminar marca",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "responses": {
          "200": { "description": "Eliminada" }
        }
      }
    },
    "/api/camisas": {
      "get": {
        "tags": ["Camisas"],
        "summary": "Listar camisas",
        "responses": {
          "200": {
            "description": "Lista",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CamisaSalida" }
                }
              }
            }
          },
          "204": { "description": "Sin datos" }
        }
      },
      "post": {
        "tags": ["Camisas"],
        "summary": "Crear camisa",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CamisaGuardar" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CamisaSalida" }
              }
            }
          }
        }
      }
    },
    "/api/camisas/paginado": {
      "get": {
        "tags": ["Camisas"],
        "summary": "Listar camisas paginadas",
        "parameters": [
          { "$ref": "#/components/parameters/page" },
          { "$ref": "#/components/parameters/size" }
        ],
        "responses": {
          "200": { "description": "Página" },
          "204": { "description": "Sin datos" }
        }
      }
    },
    "/api/camisas/buscar": {
      "get": {
        "tags": ["Camisas"],
        "summary": "Buscar camisas por nombre",
        "parameters": [
          {
            "name": "nombre",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Resultados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CamisaSalida" }
                }
              }
            }
          },
          "204": { "description": "Sin datos" }
        }
      }
    },
    "/api/camisas/categoria/{categoriaId}": {
      "get": {
        "tags": ["Camisas"],
        "summary": "Listar camisas por categoría",
        "parameters": [
          {
            "name": "categoriaId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CamisaSalida" }
                }
              }
            }
          },
          "204": { "description": "Sin datos" }
        }
      }
    },
    "/api/camisas/{id}": {
      "get": {
        "tags": ["Camisas"],
        "summary": "Obtener camisa por ID",
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "responses": {
          "200": {
            "description": "Encontrada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CamisaSalida" }
              }
            }
          },
          "404": { "description": "No existe" }
        }
      },
      "put": {
        "tags": ["Camisas"],
        "summary": "Editar camisa o stock",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CamisaModificar" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Actualizada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CamisaSalida" }
              }
            }
          },
          "404": { "description": "No existe" }
        }
      },
      "delete": {
        "tags": ["Camisas"],
        "summary": "Eliminar camisa",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "responses": {
          "200": { "description": "Eliminada" }
        }
      }
    },
...
  }
}